OPENEUROPA-1247: Move the SearchParameters and its interface to new Search folder.
